Fixed enum classes

+++ zone/client.cpp
@@ -11607,7 +11607,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 	bool passes = false;
 
 	switch (bucket_comparison) {
-		case MerchantBucketComparison::BucketEqualTo:
+		case  static_cast<uint8>(MerchantBucketComparison::BucketEqualTo):
 		{
 			if (player_value != bucket_value) {
 				break;
@@ -11616,7 +11616,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketNotEqualTo:
+		case static_cast<uint8>(MerchantBucketComparison::BucketNotEqualTo):
 		{
 			if (player_value == bucket_value) {
 				break;
@@ -11625,7 +11625,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketGreaterThanOrEqualTo:
+		case static_cast<uint8>(MerchantBucketComparison::BucketGreaterThanOrEqualTo):
 		{
 			if (player_value < bucket_value) {
 				break;
@@ -11634,7 +11634,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketLesserThanOrEqualTo:
+		case static_cast<uint8>(MerchantBucketComparison::BucketLesserThanOrEqualTo):
 		{
 			if (player_value > bucket_value) {
 				break;
@@ -11643,7 +11643,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketGreaterThan:
+		case static_cast<uint8>(MerchantBucketComparison::BucketGreaterThan):
 		{
 			if (player_value <= bucket_value) {
 				break;
@@ -11652,7 +11652,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketLesserThan:
+		case static_cast<uint8>(MerchantBucketComparison::BucketLesserThan):
 		{
 			if (player_value >= bucket_value) {
 				break;
@@ -11662,7 +11662,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 
 			break;
 		}
-		case MerchantBucketComparison::BucketIsAny:
+		case static_cast<uint8>(MerchantBucketComparison::BucketIsAny):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11683,7 +11683,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketIsNotAny:
+		case static_cast<uint8>(MerchantBucketComparison::BucketIsNotAny):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11704,7 +11704,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketIsBetween:
+		case static_cast<uint8>(MerchantBucketComparison::BucketIsBetween):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11721,7 +11721,7 @@ bool Client::CheckMerchantDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case MerchantBucketComparison::BucketIsNotBetween:
+		case static_cast<uint8>(MerchantBucketComparison::BucketIsNotBetween):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11856,7 +11856,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 	bool passes = false;
 
 	switch (bucket_comparison) {
-		case BotSpellBucketComparison::BucketEqualTo:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketEqualTo):
 		{
 			if (player_value != bucket_value) {
 				break;
@@ -11865,7 +11865,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketNotEqualTo:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketNotEqualTo):
 		{
 			if (player_value == bucket_value) {
 				break;
@@ -11874,7 +11874,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketGreaterThanOrEqualTo:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketGreaterThanOrEqualTo):
 		{
 			if (player_value < bucket_value) {
 				break;
@@ -11883,7 +11883,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketLesserThanOrEqualTo:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketLesserThanOrEqualTo):
 		{
 			if (player_value > bucket_value) {
 				break;
@@ -11892,7 +11892,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketGreaterThan:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketGreaterThan):
 		{
 			if (player_value <= bucket_value) {
 				break;
@@ -11901,7 +11901,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketLesserThan:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketLesserThan):
 		{
 			if (player_value >= bucket_value) {
 				break;
@@ -11911,7 +11911,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 
 			break;
 		}
-		case BotSpellBucketComparison::BucketIsAny:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketIsAny):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11932,7 +11932,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketIsNotAny:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketIsNotAny):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11953,7 +11953,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketIsBetween:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketIsBetween):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
@@ -11970,7 +11970,7 @@ bool Client::CheckBotSpellDataBucket(uint8 bucket_comparison, std::string bucket
 			passes = true;
 			break;
 		}
-		case BotSpellBucketComparison::BucketIsNotBetween:
+		case static_cast<uint8>(BotSpellBucketComparison::BucketIsNotBetween):
 		{
 			bucket_checks = Strings::Split(bucket_value, "|");
 			if (bucket_checks.empty()) {
